backend/requirements.txt
django==5.2.*
django-debug-toolbar==5.2.*
Pillow==11.2.*
environs[django]==14.2.*
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
requests==2.32.5
geopy==2.4.1
phonenumbers==9.0.16
rollbar==1.3.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
